move test with no visual component to the functional suite
